Foreign-Connected PAC Contributions

Every election cycle brings its own brand of excitement – and lots of money. Political donations are of particular interest to political scientists and other researchers studying politics and voting patterns. They are also of interest to citizens who want to stay informed of how much money their candidates raise and where that money comes from.

pac_2020

Format

A tibble with 224 rows and 5 variables:

name

Name of company (American divisions of a foreign company)

country_parent

Country of origin and (foreign) parent company

total

Total PAC contributions to all parties

dems

PAC contributions to the Democratic Party

repubs

PAC contributions to the Republican Party

Source

https://www.opensecrets.org/political-action-committees-pacs/foreign-connected-pacs

Details

In the United States, “only American citizens (and immigrants with green cards) can contribute to federal politics, but the American divisions of foreign companies can form political action committees (PACs) and collect contributions from their American employees.” (Source: OpenSecrets.org)

The data come from the OpenSecrets.org.

The data were scraped from the OpenSecrets.org website in June 2020. It is part of a series of similar datasets in this package consisting of datasets collected from OpenSecrets via webscraping.

This particular dataset consists of the data for foreign-connected PAC contrbutions for the 2020 election cycle, with some cleaning also done to make it more accessible in introductory data science education.

Examples

library(tidyverse)

#> ── Attaching packages ─────────────────────────────────────── tidyverse 1.3.0 ──

#> ✔ tibble  3.0.3     ✔ purrr   0.3.4
#> ✔ tidyr   1.1.1     ✔ stringr 1.4.0
#> ✔ readr   1.3.1     ✔ forcats 0.5.0

#> ── Conflicts ────────────────────────────────────────── tidyverse_conflicts() ──
#> ✖ dplyr::filter() masks stats::filter()
#> ✖ dplyr::lag()    masks stats::lag()

glimpse(pac_2020)

#> Rows: 224
#> Columns: 5
#> $ name           <chr> "7-Eleven", "ABB Group (ABB Group)", "Accenture (Accen…
#> $ country_parent <chr> "Japan/Seven & I Holdings", "Switzerland/Asea Brown Bo…
#> $ total          <chr> "$1,000", "$9,500", "$82,000", "$14,000", "$166,000", …
#> $ dems           <chr> "$0", "$3,500", "$49,000", "$5,000", "$69,000", "$25,7…
#> $ repubs         <chr> "$1,000", "$6,000", "$33,000", "$9,000", "$97,000", "$…