Instructional staff employee trends

The American Association of University Professors (AAUP) is a nonprofit membership association of faculty and other academic professionals. This dataset released by the AAUP shows trends in instructional staff employees between 1975 and 2011. Values given are percentages of full-time tenured, full-time tenure-track, full-time non-tenure-track, and part-time faculty as well as graduate student employees for each year in the dataset. Values may not add up to 100% for a given year due to rounding errors.

instructors

Format

A tibble with 11 rows and 6 variables:

year

Year

full_time_tenured

Percentage of Full-Time Tenured Faculty

full_time_tenure_track

Percentage of Full-Time Tenure-Track Faculty

full_time_non_tenure_track

Percentage of Full-Time Non-Tenure-Track Faculty

part_time

Percentage of Part-Time Faculty

grad_student

Percentage of Graduate Student Employees

Source

https://www.aaup.org/sites/default/files/files/AAUP_Report_InstrStaff-75-11_apr2013.pdf

Examples

instructors

#> # A tibble: 11 x 6
#>     year full_time_tenur… full_time_tenur… full_time_non_t… part_time
#>    <dbl>            <dbl>            <dbl>            <dbl>     <dbl>
#>  1  1975             29               16.1             10.3      24  
#>  2  1989             27.6             11.4             14.1      30.4
#>  3  1993             25               10.2             13.6      33.1
#>  4  1995             24.8              9.6             13.6      33.2
#>  5  1999             21.8              8.9             15.2      35.5
#>  6  2001             20.3              9.2             15.5      36  
#>  7  2003             19.3              8.8             15        37  
#>  8  2005             17.8              8.2             14.8      39.3
#>  9  2007             17.2              8               14.9      40.5
#> 10  2009             16.8              7.6             15.1      41.1
#> 11  2011             16.7              7.4             15.4      41.3
#> # … with 1 more variable: grad_student <dbl>