Foreign-Connected PAC Contributions, 2020, clean data

Every election cycle brings its own brand of excitement – and lots of money. Political donations are of particular interest to political scientists and other researchers studying politics and voting patterns. They are also of interest to citizens who want to stay informed of how much money their candidates raise and where that money comes from.

pac_2020

Format

A tibble with 226 rows and 6 variables:

name

Name of company (American divisions of a foreign company)

country

Country of origin

parent

PAC parent company

total

Total PAC contributions to all parties

dems

PAC contributions to the Democratic Party

repubs

PAC contributions to the Republican Party

Source

https://www.opensecrets.org/political-action-committees-pacs/foreign-connected-pacs

Details

In the United States, “only American citizens (and immigrants with green cards) can contribute to federal politics, but the American divisions of foreign companies can form political action committees (PACs) and collect contributions from their American employees.” (Source: OpenSecrets.org)

The data come from the OpenSecrets.org.

The data were scraped from the OpenSecrets.org website in June 2020. It is part of a series of similar datasets in this package consisting of datasets collected from OpenSecrets via webscraping.

This particular dataset consists of the data for foreign-connected PAC contributions for the 2020 election cycle, with some cleaning also done to make it more accessible in introductory data science education.

See also

Examples

library(tibble)

glimpse(pac_2020)

#> Rows: 226
#> Columns: 6
#> $ name    <chr> "7-Eleven", "ABB Group (ABB Group)", "Accenture (Accenture)", …
#> $ country <chr> "Japan", "Switzerland", "Ireland", "France", "Netherlands", "I…
#> $ parent  <chr> "Seven & I Holdings", "Asea Brown Boveri", "Accenture plc", "L…
#> $ total   <dbl> 2000, 15600, 82000, 24800, 176500, 89750, 66300, 500, 307500, …
#> $ dems    <dbl> 1000, 6800, 49000, 11800, 72000, 30750, 33150, 0, 171500, 4450…
#> $ repubs  <dbl> 1000, 8800, 33000, 13000, 104500, 59000, 33150, 500, 136000, 4…