Foreign-Connected PAC Contributions

Every election cycle brings its own brand of excitement – and lots of money. Political donations are of particular interest to political scientists and other researchers studying politics and voting patterns. They are also of interest to citizens who want to stay informed of how much money their candidates raise and where that money comes from.

pac_2020_raw

Format

A tibble with 224 rows and 5 variables:

PAC Name (Affiliate)

Name of company (American divisions of a foreign company)

Country of Origin/Parent Company

Country of origin and (foreign) parent company

Total

Total PAC contributions to all parties

Dems

PAC contributions to the Democratic Party

Repubs

PAC contributions to the Republican Party

Source

https://www.opensecrets.org/political-action-committees-pacs/foreign-connected-pacs

Details

In the United States, “only American citizens (and immigrants with green cards) can contribute to federal politics, but the American divisions of foreign companies can form political action committees (PACs) and collect contributions from their American employees.” (source: OpenSecrets.org)

The data come from the OpenSecrets.org.

The data were scraped from the OpenSecrets.org website in June 2020. It is part of a series of similar datasets in this package consisting of datasets collected from OpenSecrets via webscraping.

This particular dataset consists of the data for foreign-connected PAC contrbutions for the 2020 election cycle, in raw form as it was when scraped from the website

Examples

pac_2020_raw

#> # A tibble: 224 x 5
#>    `PAC Name (Affiliate)`            `Country of Origin/Pa… Total  Dems   Repubs
#>    <chr>                             <chr>                  <chr>  <chr>  <chr> 
#>  1 "7-Eleven\n                     … Japan/Seven & I Holdi… $1,000 $0     $1,000
#>  2 "ABB Group\n                    … Switzerland/Asea Brow… $9,500 $3,500 $6,000
#>  3 "Accenture\n                    … Ireland/Accenture plc  $82,0… $49,0… $33,0…
#>  4 "Air Liquide America\n          … France/L'Air Liquide … $14,0… $5,000 $9,000
#>  5 "Airbus Group\n                 … Netherlands/Airbus Gr… $166,… $69,0… $97,0…
#>  6 "Alkermes Inc\n                 … Ireland/Alkermes Plc   $77,2… $25,7… $51,5…
#>  7 "Allianz of America\n           … Germany/Allianz AG Ho… $54,5… $25,3… $29,1…
#>  8 "Anheuser-Busch\n               … Belgium/Anheuser-Busc… $256,… $129,… $126,…
#>  9 "AON Corp\n                     … UK/AON PLC             $58,0… $27,0… $31,0…
#> 10 "APL Maritime\n                 … France/CMA CGM SA      $16,0… $10,5… $5,500
#> # … with 214 more rows