Foreign-Connected PAC Contributions, 2020, raw data

Every election cycle brings its own brand of excitement – and lots of money. Political donations are of particular interest to political scientists and other researchers studying politics and voting patterns. They are also of interest to citizens who want to stay informed of how much money their candidates raise and where that money comes from.

pac_2020_raw

Format

A tibble with 226 rows and 5 variables:

PAC Name (Affiliate)

Name of company (American divisions of a foreign company)

Country of Origin/Parent Company

Country of origin and (foreign) parent company

Total

Total PAC contributions to all parties

Dems

PAC contributions to the Democratic Party

Repubs

PAC contributions to the Republican Party

Source

https://www.opensecrets.org/political-action-committees-pacs/foreign-connected-pacs

Details

In the United States, “only American citizens (and immigrants with green cards) can contribute to federal politics, but the American divisions of foreign companies can form political action committees (PACs) and collect contributions from their American employees.” (source: OpenSecrets.org)

The data come from the OpenSecrets.org.

The data were scraped from the OpenSecrets.org website in June 2020. It is part of a series of similar datasets in this package consisting of datasets collected from OpenSecrets via webscraping.

This particular dataset consists of the data for foreign-connected PAC contributions for the 2020 election cycle, in raw form as it was when scraped from the website on 11 Nov 2020.

See also

Examples

pac_2020_raw
#> # A tibble: 226 × 5
#>    `PAC Name (Affiliate)`                `Country of Origin…` Total Dems  Repubs
#>    <chr>                                 <chr>                <chr> <chr> <chr> 
#>  1 7-Eleven                              Japan/Seven & I Hol… $2,0… $1,0… $1,000
#>  2 ABB Group (ABB Group)                 Switzerland/Asea Br… $15,… $6,8… $8,800
#>  3 Accenture (Accenture)                 Ireland/Accenture p… $82,… $49,… $33,0…
#>  4 Air Liquide America                   France/L'Air Liquid… $24,… $11,… $13,0…
#>  5 Airbus Group                          Netherlands/Airbus … $176… $72,… $104,…
#>  6 Alkermes Inc                          Ireland/Alkermes Plc $89,… $30,… $59,0…
#>  7 Allianz of America (Allianz)          Germany/Allianz AG … $66,… $33,… $33,1…
#>  8 AMG Vanadium                          Netherlands/AMG Adv… $500  $0    $500  
#>  9 Anheuser-Busch (Anheuser-Busch InBev) Belgium/Anheuser-Bu… $307… $171… $136,…
#> 10 AON Corp (AON plc)                    UK/AON PLC           $87,… $44,… $43,0…
#> # … with 216 more rows