Foreign-Connected PAC Contributions

Every election cycle brings its own brand of excitement – and lots of money. Political donations are of particular interest to political scientists and other researchers studying politics and voting patterns. They are also of interest to citizens who want to stay informed of how much money their candidates raise and where that money comes from.

pac_all

Format

A tibble with 2245 rows and 6 variables:

name

Name of company (American divisions of a foreign company)

country_parent

Country of origin and (foreign) parent company

total

Total PAC contributions to all parties

dems

PAC contributions to the Democratic Party

repubs

PAC contributions to the Republican Party

year

Election cycle year

Source

https://www.opensecrets.org/political-action-committees-pacs/foreign-connected-pacs

Details

In the United States, “only American citizens (and immigrants with green cards) can contribute to federal politics, but the American divisions of foreign companies can form political action committees (PACs) and collect contributions from their American employees.” (source: OpenSecrets.org)

The data come from the OpenSecrets.org.

The data were scraped from the OpenSecrets.org website in June 2020. It is part of a series of similar datasets in this package consisting of datasets collected from OpenSecrets via webscraping.

This particular dataset consists of the data for foreign-connected PAC contrbutions for all available elections cycles, with some minor cleaning also done to make it more accessible in introductory data science education.

Examples

library(tidyverse)

glimpse(pac_all)

#> Rows: 2,245
#> Columns: 6
#> $ name           <chr> "ABB Daimler-Benz Transportation", "ABB Group", "AE St…
#> $ country_parent <chr> "Germany/Daimler-Benz AG", "Switzerland/Asea Brown Bov…
#> $ total          <chr> "$15,775", "$48,500", "$30,500", "$41,700", "$31,000",…
#> $ dems           <chr> "$10,775", "$19,750", "$3,500", "$10,100", "$8,500", "…
#> $ repubs         <chr> "$5,000", "$28,750", "$27,000", "$31,600", "$22,500", …
#> $ year           <dbl> 1998, 1998, 1998, 1998, 1998, 1998, 1998, 1998, 1998, …