Student survey

During the first week of class in Spring 2015, students in an introductory statistics course at Duke University came up with survey questions designed to collect data on variables they are interested in working with throughout the semester. Students answered these questions anonymously in an online survey, and the resulting data are distributed here. Note that these data are somewhat cleaned and potentially identifiable cases are removed.

student_survey

Format

A tibble with 80 rows and 53 variables:

pets

How many pets do you have? Enter a positive whole number

us_region

What part of the US are you from? (New England, Mid-Atlantic, Midwest, Southwest, Southeast, California, Pacific Northwest, Not from the US)

class_year

What is your class year? (First-year, Sophomore, Junior, Senior)

hair_color

What color hair do you have? (Brown, Blonde, Black)

campus

What campus do you live on? (East, West, Central, Off campus)

watch_sports

What is your favorite sport to watch? (Basketball, Football, Soccer, Other, I dont watch sports)

beyonce_love

DO YOU LOVE BEYONCE? (hell yes, yes, meh, no, I don't care. \#teamnicki)

fav_artist

Who is your favorite artist? (Beyonce, Taylor swift, Kanye West, Kendrick Lamar, Florida-Georgia Line, none of these)

social_network

Favorite social network? (Facebook, Twitter, Instagram, Yikyak, Vine, Snapchat, Youtube, I don't use social media)

relationship_status

Are you currently in a relationship? (yes, no, it's complicated)

num_siblings

How many siblings do you have? Enter a positive whole number. If you are an only child, enter 0

num_languages

How many languages can you carry a conversation in? Enter a number 1 or above, since you at least speak your native language fluently

nights_drinking

How many times a week do you drink, on average?

pbj_or_n

Nutella, Peanut Butter, or Jelly? (Nutella, Peanut butter, Jelly, Some combination of two, All three, None)

tenting

What level of tenting are you doing? (none, black, blue, white, unsure)

countries_visited

How many countries have you visited excluding your home country?

first_kiss

How old were you when you had your first kiss? If you have yet to have your first kiss, enter NA. Otherwise, enter your age in years when you had your first kiss

hours_watching_television

How many hours of television do you watch per week (includes Netflix and online streaming)?

perimeter

Inside or Outside the Perimeter? (Inside, Outside, What even is the Perimeter, All of the Above)

fav_campus_eatery

What is your favorite on-campus eatery? (The Loop, Divinity Refectory, Edens Cafe (lol), Blue Express, Other, None of the above)

fb_visits_per_day

How many times do you go on Facebook per day? Enter a positive whole number. If you never visit Facebook, enter 0

ac

Does your house/apt/dorm have AC? (Yes, No)

condiment

If you were a condiment what would you be? (Soy Sauce, Ketchup, Sriracha, Honey Mustard, Maple Syrup)

vegetarian

Do you consider yourself a vegetarian? (Yes, No)

voted_for_president

Did you vote in last year's presidential election? (yes; no, I was not eligible due to age or citizenship status; no, I was eligible but chose not to vote)

social_club

What kind of social group do you belong to? (Greek life, SLG, Athletic Team, Independent)

space_time

Would you rather travel through outer space or time? (Outer space, Time)

university_applications

How many universities did you apply to? Enter a positive whole number 1 or above

pizza_consumption

How many pieces of pizza do you eat per week on average? Enter a positive whole number. If you never eat pizza, enter 0

sick

How many days were you sick in the past month? Enter a positive whole number. If you were never sick in the past month, enter 0

games_attended

How many basketball games did you attend in 2014? Enter a positive whole number. If you didn't attend any basketball games in 2014, enter 0

pepsi_or_coke

Pepsi or Coke? (Pepsi, Coke, No preference, Neither they're nasty)

fav_friend_character

Who is your favorite Friend? (Joey, Chandler, Monica, Rachel, Phoebe, Ross, Janice, Gunther)

continents_visited

How many continents have you been to?

game_of_thrones

Who shall rule the Realm? (House Targaryen, House Tyrell, House Stark, House Lannister, House Martell, House Greyjoy, I don't know what you are talking about)

netflix_binge_show

What is your Netflix binge show? (House of Cards, Gilmore Girls, Breaking Bad, How I Met Your Mother, Friends, Battlestar Galactica)

chipotle_order

What do you order at Chipotle? (Burrito Bowl, Quesadilla, Salad, Margarita (no food), Hard tacos, Soft tacos, Burrito)

fav_pokemon

What is your favorite pokemon? (Charizard, Blastoise, Venusaur, Jigglypuff, Rayquaza)

first_tooth

When did you lose your first tooth?

fav_cheese

What is your favorite cheese? (Blue, Goat, Feta, Cheddar, Pepperjack, Brie, None of the above)

cat_or_dog

Cat or Dog? (Cat, Dog, I'm not an animal person)

fav_late_night_food

What is your late night food of choice? (Pizza, Jimmy John's, Cosmic, Cookout, Pitchforks)

fav_dessert_flavor

What your favorite dessert flavor? (chocolate, vanilla, fruity, sour, sweet-and-salty, other, I don't eat dessert because I don't like fun)

fav_off_campus_restaurant

What is your favorite off-campus Restaurant? (Satisfactions, Cosmic, Harris Teeter's, Parker and Otis, Chamas, Chipotle)

vending_machine

Do you raid the vending machine at night? (Yes, No, Sometimes, Shhhhh...)

multicultural

Do you consider yourself multicultural? (Yes, No)

smell

Which smell do you prefer? (Pine, Clean Laundry, Pizza, Bakery, Old Spice/Axe/Perfume, Sharpie)

parents_age

What is the average age of your parents?

marketplace_worst_meal

What is Marketplace's worst meal? (chocolate covered crickets, salad bar, powdered eggs, the giant roasted hog, watery pasta)

best_book_turned_show

Favorite book series turned Movie or TV Show? (Harry Potter, Game of Thrones, Lord of the Rings-Hobbit, Hunger Games, Twilight)

fav_fantasy_universe

Favorite fantasy universe? (Legend of Zelda, Game of Thrones, Lord of the Rings, Harry Potter)

hogwarts_house

Which Harry Potter house do you belong to? (Gryffindor (because I'm a badass), Ravenclaw (because I'm better than you), Hufflepuff (because I'm good at something), Slytherin (because I'll cut you))

shooters

How often do you go to Shooters a month? Enter a positive whole number. Enter 0 if you never go to Shooters

Examples

student_survey

#> # A tibble: 80 x 53
#>     pets us_region class_year hair_color campus watch_sports beyonce_love
#>    <int> <chr>     <chr>      <chr>      <chr>  <chr>        <chr>       
#>  1     1 Southeast Junior     Brown      West   Basketball   yes         
#>  2     0 Southeast First-year Brown      East   Football     hell yes    
#>  3     0 Not from… Sophomore  Black      West   Soccer       no          
#>  4     1 Pacific … Junior     Brown      Off c… Football     hell yes    
#>  5     7 Southwest Sophomore  Brown      West   I don't wat… hell yes    
#>  6     1 Mid-Atla… Sophomore  Brown      West   Basketball   hell yes    
#>  7     3 Midwest   First-year Blonde     East   I don't wat… meh         
#>  8     0 Mid-Atla… Sophomore  Brown      West   Basketball   yes         
#>  9     1 Mid-Atla… Sophomore  Blonde     Centr… Basketball   hell yes    
#> 10     1 Midwest   First-year Brown      East   Basketball   yes         
#> # … with 70 more rows, and 46 more variables: fav_artist <chr>,
#> #   social_network <chr>, relationship_status <chr>, num_siblings <int>,
#> #   num_languages <chr>, nights_drinking <dbl>, pbj_or_n <chr>, tenting <chr>,
#> #   countries_visited <int>, first_kiss <int>, hours_watching_television <dbl>,
#> #   perimeter <chr>, fav_campus_eatery <chr>, fb_visits_per_day <int>,
#> #   ac <chr>, condiment <chr>, vegetarian <chr>, voted_for_president <chr>,
#> #   social_club <chr>, space_time <chr>, university_applications <int>,
#> #   pizza_consumption <dbl>, sick <int>, games_attended <int>,
#> #   pepsi_or_coke <chr>, fav_friend_character <chr>, continents_visited <int>,
#> #   game_of_thrones <chr>, netflix_binge_show <chr>, chipotle_order <chr>,
#> #   fav_pokemon <chr>, first_tooth <int>, fav_cheese <chr>, cat_or_dog <chr>,
#> #   fav_late_night_food <chr>, fav_dessert_flavor <chr>,
#> #   fav_off_campus_restaurant <chr>, vending_machine <chr>,
#> #   multicultural <chr>, smell <chr>, parents_age <dbl>,
#> #   marketplace_worst_meal <chr>, best_book_turned_show <chr>,
#> #   fav_fantasy_universe <chr>, hogwarts_house <chr>, shooters <int>